requirements.txt
# Core LLM & Agent Framework
langchain==0.3.25
langchain-core==0.3.60
langgraph==0.2.55
langchain-openai==0.2.7
langchain-community==0.3.21

# Vector DB & RAG
chromadb==0.5.23

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Utilities
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.9.2
tiktoken==0.7.0
